Add Support For PSP isos
Extend the program to also support Playstation Portable ISO files. For
this file type, the SFO file is located in a different subdirectory. Use
isoinfo to list the directory structure and locate the SFO file. This
also gives a more descriptive exception if a non-playstation image is
loaded with this program.

Playstation Portable SFO files contain a variety of additional tags in
the PARAM.SFO file. Adding all known options from the specified source
here.